+++ go.mod
@@ -3,6 +3,7 @@ module github.com/shomali11/slacker
 go 1.14
 
 require (
+	github.com/robfig/cron v1.2.0 // indirect
 	github.com/shomali11/commander v0.0.0-20220716022157-b5248c76541a
 	github.com/shomali11/proper v0.0.0-20180607004733-233a9a872c30
 	github.com/slack-go/slack v0.12.1
